Refactor service package structure and add review entity with related functionality
