--- nukkadseva-api-docs.yaml
@@ -1,134 +1,0 @@
-openapi: 3.0.1
-info:
-  title: OpenAPI definition
-  version: v0
-servers:
-- url: http://localhost:8080
-  description: Generated server url
-paths:
-  /verify-otp:
-    post:
-      tags:
-      - user-controller
-      operationId: verifyOtp
-      requestBody:
-        content:
-          application/json:
-            schema:
-              $ref: "#/components/schemas/VerifyOtpRequest"
-        required: true
-      responses:
-        "200":
-          description: OK
-          content:
-            '*/*':
-              schema:
-                $ref: "#/components/schemas/ApiResponse"
-  /send-verification-otp:
-    post:
-      tags:
-      - user-controller
-      operationId: sendVerificationOtp
-      parameters:
-      - name: email
-        in: query
-        required: true
-        schema:
-          type: string
-      responses:
-        "200":
-          description: OK
-          content:
-            '*/*':
-              schema:
-                $ref: "#/components/schemas/OtpTokenResponse"
-  /customer/register:
-    post:
-      tags:
-      - user-controller
-      operationId: customerRegistration
-      requestBody:
-        content:
-          application/json:
-            schema:
-              $ref: "#/components/schemas/UserRequest"
-        required: true
-      responses:
-        "200":
-          description: OK
-          content:
-            '*/*':
-              schema:
-                type: string
-  /customer/login:
-    post:
-      tags:
-      - user-controller
-      operationId: customerLogin
-      requestBody:
-        content:
-          application/json:
-            schema:
-              $ref: "#/components/schemas/UserRequest"
-        required: true
-      responses:
-        "200":
-          description: OK
-          content:
-            '*/*':
-              schema:
-                $ref: "#/components/schemas/AuthResponse"
-  /:
-    get:
-      tags:
-      - user-controller
-      operationId: welcom_1
-      responses:
-        "200":
-          description: OK
-          content:
-            '*/*':
-              schema:
-                type: string
-components:
-  schemas:
-    VerifyOtpRequest:
-      type: object
-      properties:
-        otp:
-          type: string
-        token:
-          type: string
-    ApiResponse:
-      type: object
-      properties:
-        statusCode:
-          type: string
-        message:
-          type: string
-    OtpTokenResponse:
-      type: object
-      properties:
-        status:
-          type: string
-        message:
-          type: string
-        token:
-          type: string
-    UserRequest:
-      required:
-      - email
-      - password
-      type: object
-      properties:
-        email:
-          type: string
-        password:
-          type: string
-    AuthResponse:
-      type: object
-      properties:
-        token:
-          type: string
-        email:
-          type: string
