+++ src/main/java/com/nukkadseva/nukkadsevabackend/config/MailConfig.java
@@ -1,5 +1,6 @@
 package com.nukkadseva.nukkadsevabackend.config;
 
+import io.github.cdimascio.dotenv.Dotenv;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.mail.javamail.JavaMailSender;
@@ -10,14 +11,18 @@ import java.util.Properties;
 @Configuration
 public class MailConfig {
 
+    Dotenv dotenv = Dotenv.load();
+    String host = dotenv.get("MAIL_SERVER");
+    String username = dotenv.get("MAIL_USERNAME");
+    String password = dotenv.get("MAIL_PASSWORD");
+
     @Bean
     public JavaMailSender javaMailSender() {
         JavaMailSenderImpl mailSender = new JavaMailSenderImpl();
-        mailSender.setHost("smtp.gmail.com");
+        mailSender.setHost(host);
         mailSender.setPort(587);
-        mailSender.setUsername("yjamal710@gmail.com");
-        mailSender.setPassword("raju vysl sfee aoxu"); //TODO: use environment variables
-
+        mailSender.setUsername(username);
+        mailSender.setPassword(password);
         Properties props = new Properties();
         props.put("mail.transport.protocol", "smtp");
         props.put("mail.smtp.auth", "true");
