+++ src/main/java/com/nukkadseva/nukkadsevabackend/entity/Review.java
@@ -1,0 +1,58 @@
+package com.nukkadseva.nukkadsevabackend.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.Column;
+import jakarta.persistence.Table;
+import jakarta.persistence.Index;
+import jakarta.persistence.Id;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.FetchType;
+import jakarta.validation.constraints.Max;
+import jakarta.validation.constraints.Min;
+import lombok.Getter;
+import lombok.Setter;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+import java.time.LocalDateTime;
+
+@Getter
+@Setter
+@Entity
+// Add indexes for foreign keys for better query performance
+@Table(name = "reviews", indexes = {
+        @Index(name = "idx_reviews_customer_id", columnList = "customer_id"),
+        @Index(name = "idx_reviews_provider_id", columnList = "provider_id")
+})
+public class Review {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "review_id")
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "customer_id", nullable = false)
+    private Customers customer;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "provider_id", nullable = false)
+    private Provider provider;
+
+    @Min(1)
+    @Max(5)
+    @Column(nullable = false)
+    private byte rating;
+
+    @Column(columnDefinition = "TEXT")
+    private String comment;
+
+    @CreationTimestamp
+    @Column(name = "created_at", updatable = false)
+    private LocalDateTime createdAt;
+
+    @UpdateTimestamp
+    @Column(name = "updated_at")
+    private LocalDateTime updatedAt;
+}
