+++ src/main/java/com/nukkadseva/nukkadsevabackend/service/AuthenticationFacade.java
@@ -1,4 +1,4 @@
-package com.nukkadseva.nukkadsevabackend.service.userservice;
+package com.nukkadseva.nukkadsevabackend.service;
 
 import org.springframework.security.core.Authentication;
