+++ src/main/java/com/nukkadseva/nukkadsevabackend/util/AuthenticationFacadeService.java
@@ -1,12 +1,11 @@
-package com.nukkadseva.nukkadsevabackend.service.implementation;
+package com.nukkadseva.nukkadsevabackend.util;
 
-import com.nukkadseva.nukkadsevabackend.service.userservice.AuthenticationFacade;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 
 @Component
-public class AuthenticationFacadeImpl implements AuthenticationFacade {
+public class AuthenticationFacadeService implements com.nukkadseva.nukkadsevabackend.service.AuthenticationFacade {
     @Override
     public Authentication getAuthenticationFacade() {
         return SecurityContextHolder.getContext().getAuthentication();
